Fix state changing on removal of values
