+++ packages/components/src/Form/Inputs/Combobox/utils/useInputEvents.ts
@@ -43,6 +43,20 @@ import { ComboboxActionType, ComboboxState } from './state'
 import { useBlur } from './useBlur'
 import { useKeyDown } from './useKeyDown'
 
+// Walks up the dom tree from an element to a containingAncestor checking for a button
+// Used for determining whether a mousedown/click should toggle the option list
+function checkForButton(
+  element: Element,
+  containingAncestor: Element
+): boolean {
+  if (element === containingAncestor) return false
+  if (!element.parentElement) return false
+  if (element.tagName === 'BUTTON') {
+    return true
+  }
+  return checkForButton(element.parentElement, containingAncestor)
+}
+
 export function useInputEvents<
   TProps extends
     | ComboboxInputProps
@@ -112,6 +126,12 @@ export function useInputEvents<
 
   const handleMouseDownClick = useCallback(
     (e: ReactMouseEvent<HTMLElement>) => {
+      // Without this, when clicking on a "clear" or "remove value" icon button
+      // the list will flash open & closed (if closed)
+      // or unnecessarily close (if open)
+      if (checkForButton(e.target as Element, e.currentTarget as Element)) {
+        return
+      }
       if (state === ComboboxState.IDLE) {
         // Opening a closed list
         transition &&
