+++ packages/components/src/Form/Inputs/Select/index.ts
@@ -26,5 +26,4 @@
 
 export * from './Select'
 export * from './SelectMulti'
-export * from './SelectMultiCreateOption'
 export * from './SelectOptions'
