+++ packages/www/src/documentation/components/forms/select-multi.mdx
@@ -8,10 +8,9 @@ propsOf: SelectMulti
 The `SelectMulti` component is an extension of the [`Select`](/components/forms/select) component that supports multiple selected values and has the following specific features:
 
 - The input field renders an underlying [`InputChips`](/components/forms/input-chips)
-- An on-the-fly "Create" option allows the user to enter custom values (see `showCreate` prop)
-- Option list stays open by default so the user can select multiple options (see `closeOnSelect` prop)
+- An on-the-fly "Create" option allows the user to enter custom values (see [`showCreate`](#showcreate) prop)
+- Option list stays open by default so the user can select multiple options (see [`closeOnSelect`](#closeonselect) prop)
 - Values can be cleared individually or all at once (does not use the `Select` component's `isClearable` prop)
-- Aside from `isClearable`, all features of `Select` are supported in `SelectMulti`
 
 ## Basic
 
@@ -32,14 +31,13 @@ The `SelectMulti` component is an extension of the [`Select`](/components/forms/
 />
 ```
 
-## Filtering
+## showCreate
 
-The `isFilterable` prop allows the user to type in the input, triggering the `onFilter` callback,
-which should be used to narrow the options passed to the `SelectMulti`. Use the `showCreate` prop to allow
-the user to free-form input values.
+As with `Select`, the `isFilterable` prop allows the user to type in the input, triggering the `onFilter` callback,
+which should be used to narrow the options passed to the `SelectMulti`.
 
-When `isFilterable` is true, use the `showCreate` prop along with the
-`formatCreateLabel` (defaults to "Create \_\_") and `createOptionPosition` (defaults to last) to allow the user to free-form values.
+When `isFilterable` is true, use the `showCreate` prop along with the `formatCreateLabel`(defaults to "Create \_\_")
+to allow the user to free-form values.
 
 ```jsx
 ;() => {
@@ -86,4 +84,26 @@ When `isFilterable` is true, use the `showCreate` prop along with the
     />
   )
 }
+```
+
+## closeOnSelect
+
+With the `closeOnSelect` prop, the option list closes after an option is selected.
+
+```jsx
+<SelectMulti
+  options={[
+    { value: 'Cheddar' },
+    { value: 'Gouda' },
+    { value: 'Swiss' },
+    { value: 'Feta' },
+    { value: 'Mascarpone' },
+    { value: 'Brie' },
+    { value: 'Mozzarella' },
+    { value: 'Cotija' },
+    { value: 'Pepperjack' },
+  ]}
+  defaultValues={['Swiss', 'Brie']}
+  closeOnSelect
+/>
 ```
